refactor: move Validation protocol to protocols folder
