+++ src/main/factories/login/LoginValidationCompositeFactory.spec.ts
@@ -1,6 +1,6 @@
 import ValidationComposite from '../../../presentation/helpers/validators/ValidationComposite'
 import RequiredFieldValidation from '../../../presentation/helpers/validators/RequiredFieldValidation'
-import Validation from '../../../presentation/helpers/validators/Validation'
+import Validation from '../../../presentation/protocols/Validation'
 import EmailValidation from '../../../presentation/helpers/validators/EmailValidation'
 import EmailValidator from '../../../presentation/protocols/EmailValidator'
 import LoginValidationCompositeFactory from './LoginValidationCompositeFactory'
