+++ src/main/factories/login/LoginValidationCompositeFactory.ts
@@ -1,4 +1,4 @@
-import Validation from '../../../presentation/helpers/validators/Validation'
+import Validation from '../../../presentation/protocols/Validation'
 import ValidationComposite from '../../../presentation/helpers/validators/ValidationComposite'
 import RequiredFieldValidation from '../../../presentation/helpers/validators/RequiredFieldValidation'
 import EmailValidation from '../../../presentation/helpers/validators/EmailValidation'
