+++ src/presentation/controllers/signup/SignUpController.spec.ts
@@ -2,7 +2,7 @@ import SignUpController from './SignUpController'
 import ServerError from '../../errors/ServerError'
 import { AddAccount, AddAccountModel } from '../../../domain/usecases/AddAccount'
 import AccountModel from '../../../domain/models/Account'
-import Validation from '../../helpers/validators/Validation'
+import Validation from '../../protocols/Validation'
 import HttpRequest from '../../protocols/HttpRequest'
 
 interface SutTypes {
