+++ src/presentation/helpers/validators/ValidationComposite.ts
@@ -1,4 +1,4 @@
-import Validation from './Validation'
+import Validation from '../../protocols/Validation'
 
 export default class ValidationComposite implements Validation {
   constructor (private readonly validations: Validation[]) {}
